Checkpoint with new approach, needs v9 -> v8 tests
Does not yet have the extra check to ensure v9 can't expose new
resources by putting accounts from one tx and assets from another into
static arrays of a called v8 app.
